+++ homeassistant/components/tuya/tuya_device_handlers/devices/wk/wk_IAYz2WK1th0cMLmL.py
@@ -4,19 +4,21 @@ from __future__ import annotations
 
 from ... import TUYA_QUIRKS_REGISTRY, TuyaDeviceQuirk
 from ...const import TuyaDeviceCategory, TuyaDPCode
-from ...conversion import scale_value_back_fixed_scale_1, scale_value_fixed_scale_1
 from ...homeassistant import TuyaClimateHVACMode, TuyaEntityCategory
 
 (
-    # This model has invalid scale 0 for temperature dps - force scale 1
     TuyaDeviceQuirk()
     .applies_to(category=TuyaDeviceCategory.WK, product_id="IAYz2WK1th0cMLmL")
     .add_climate(
         key="wk",
         switch_only_hvac_mode=TuyaClimateHVACMode.HEAT_COOL,
-        current_temperature_state_conversion=scale_value_fixed_scale_1,
-        target_temperature_state_conversion=scale_value_fixed_scale_1,
-        target_temperature_command_conversion=scale_value_back_fixed_scale_1,
+        current_temperature_dp_code=TuyaDPCode.UPPER_TEMP,
+        # UPPER_TEMP uses incorrect scale 1 / step 5 - convert to proper temperature
+        current_temperature_state_conversion=lambda _device, _def, value: value / 2,
+        # TEMP_SET uses incorrect scale 0 / step 5 - convert to proper temperature
+        target_temperature_dp_code=TuyaDPCode.TEMP_SET,
+        target_temperature_state_conversion=lambda _device, _def, value: value / 2,
+        target_temperature_command_conversion=lambda _device, _def, value: value * 2,
     )
     .add_switch(
         key=TuyaDPCode.CHILD_LOCK,
