+++ homeassistant/components/tuya/tuya_device_handlers/utils.py
@@ -5,23 +5,6 @@ from __future__ import annotations
 from enum import StrEnum
 
 
-def scale_value(value: int, step: float, scale: float) -> float:
-    """Official scaling function from Tuya.
-
-    See https://support.tuya.com/en/help/_detail/Kadi66s463e2q
-    """
-    return step * value / (10**scale)
-
-
-def scale_value_back(value: float, step: float, scale: float) -> int:
-    """Official scaling function from Tuya.
-
-    See https://support.tuya.com/en/help/_detail/Kadi66s463e2q
-    """
-
-    return int(value * (10**scale) / step)
-
-
 def parse_enum[T: StrEnum](enum_class: type[T], value: str | None) -> T | None:
     """Parse a string to an enum member.
